Decoupling scene creation from resource loading
Previus we would load resources pre-emptively, then create the scene tree as well as the vertex buffers. So a lot of things needed to happens before and while creating the scene manager. The changes here decouple this process so that we can create the scene tree and dynamically load resources. This is pre work for allowing to dynamically add node in the scene editor with resources.