Adding key prop to Error component and disabling loading state on error
Adding key prop to the error component so that react doesn't complain about duplicate `key`s, which are annoying and hard to diagnose. Also disabled the loading state when there is an unhandled error so that the error is render correctly. Previously, the loading state could very potentially block parts of the error.