Refactored split panel to introduce better mechanics for creating them
Adding helper function `createSplitPanel` to make it simpler to create the split panel that we use to render the canvas and the scene tree. Also reworked the split panel css and added a more semantic HTML structure so that we can create split panel with vertical as well as horizontal dividers. With the way the html and css is now structured we can create nested split panels.  From this, I have introduced a new event `panel:resize` that is triggered on the split panel element any time the resizer is moved. This helps faciliate handling/updates that need to happen when the panels are resized.