Big refactor of the different factories for scene manager, state manager, and scene udpater
The changes here are for splitting up the creation of the state manager, the scene manager, the scene updater, and the scene renderer. The goal is to make it more clear what are the different moving parts and where and when they are used. I ahve added documentation about each part to make it more clear what their porpuse is because they all feed into each other in very specific ways.