Renamed color-channels-properties and coordinate-properties components
Renaming color-channels-properties.jsx to color-channels.jsx and coordinate-properties.jsx to coordinates.jsx to make it easier to know that those components are not a properties panel.  They are actual fields rendered for different properties. Previously, it was hard to know why those two components didnt have the `node-properties` class names and why they were fragments. Hopefully now it is more clear.