Adding naive/brute-force updating of scene tree jsx
I have added a temporary way to trigger updates to re-render app.jsx that can be used whenever there are updates to scene nodes that affect siblings in the component tree. This is brute force approach to at least have a way to propagate changes, but a more principled way will be implemented in a follow up change where we trigger changes in the state manager itself that react components can subscribe to instead. But that will require a bit more work so we are adding this temporary but functional brute force udpating of the component tree.