Adding better width in properties component grid system for labels in the scene tree
Labels in different properties components would have an increasing amount of space between them and the value for the property as the right panel that hosts the scene tree itself would be resized to larger widths.  The tweaks in the CSS I am making will keep the space between the label and the value much more consistent as the scene tree panel is resized.