Adding ability to select a file for a model resource
Im in the proces of adding support for loading a model file for rendering for a static-mesh or light scene node from the scene tree information pane. This wires up the button to open up the file selector dialog (only single selection), the event handler to be able to create the URL object to load the model file, and proper rendering of just the file name because when we select a file from the file dialog all we get back is the file name, but processing the scene config you get the full path. So we are only rendering the file name, which also has the added benefit that it will nicely fit in the scene tree component where we render the file name for static-mesh and light nodes.

A follow up set of changes will refactor the resource loaded so that it can take a file path (URL) and a file name, load the file path, and update the scene node accordingly.