Wiring up the ability to select different obj files in static-mesh and light scene nodes
There are several changes in here in order to enable the ability to dynamically load obj files in scene nodes of type static-mesh and light. I updated how resources are loaded when creating a scene graph so that we use the configuration object instead of the state manager. This makes it more clear that we are really initializing everything from the scene graph configuration. Refactored the scene manager builder to create the scene manager and state manager since they are both quite tightly couple couple for most purposes. So the new createSceneGraph will instantiate the state manager as well as the scene manager. I have then introduced a resource loaded that will take in a scene node, url, and filename in order to load OBJ files into scee nodes that support rendering geometry such as static-mesh and light scene nodes.

An important change here is that when loading models from the scene tree components, the URLs are based on the file selector HTML functionality which we create URL from blobs. When this happens, we don't really have a full path to the file being loaded but rather some opaque URL. We also do get the filename being loaded, so we use that to render the resource for static-mesh and light nodes. Because of adding the need to dynamically select OBJ files, I have refactored the scene graph initialization stage to also use the node, url, and filename to more smoothly integrate that logic with dynamically loading OBJ files using the URL.createObjectURL API.