Adding tree traversal that returns all the matches for a given delegate matcher
We had code that was very boilerplate that traverses a scene tree collecting all the nodes with specific types. I have abstracted that boilerplate into a tee traversal function that takes in a matcher and returns all the items that the matcher returns true for. This new traversal functionality is used for gathering all the nodes that have resources that need to be loaded.