Adding the ability to lazy load shader programs into a scene graph
To decouple the creation of a scene graph tree, I have moved the step of creating the shader programs into its own step that happens once a scene tree is created. This will eventually also allow us to change shader programs on the fly via the editor and also allow us to shader more code in the examples being created in follow up work.

As part of this refactor, I have added a couple of helper functions for `isLight` and `isStaticMesh` so that we don't have the node type string hard coded everywhere!  In the future I will be converting these to const strings that scene-config file can use so that they don't have hard coded string in them for known node types.