Adding 3D axis HUD to show rotation of the scene
Im adding the small display of axis that tracks the rotation of the scene. This is using the an orthographic projection so that depth and angle dont distort the HUD. This also uses ambient colors so that scene lighting doesn't affect the axis HUD.

The axis are broken as
X which is of red color. Points right
Y which is of green color. Points up
Z which is of blue color. Points towards you (positive direction).