Adding perspective and orthographic panels in the scene graph tree component
After adding support for configuring projection matrices in the scene configuration, Im now adding the ability to view/update that data in the scene graph view with the addition of a perspective and an orthographic properties panel.  To make this work correctly, I have had to change the scene graph configuration to wrap the projection values in a projections object in the scene configuration. This avoid subtleties for updating the scene state manager from the scene graph viewer.