Rewrite of how the scene graph UI is created and updated
Previously, we used to build the scene graph UI from the state manager which worked ok for a lil while.  But it is more semantically correct to build it from the scene graph manager itself since that's what we are intending to render.  In order to do that, the scene tree UI will iterate thru the scene manager nodes and for any updates it uses the state manager. To reduce boilerplate code for reading and update the state of a particular node I have introduce a new component called `WithNodeState` which provides two functions; 1. getNodeState, 2. updateNodeState. That makes it quite a bit simpler to update the state of any particular node in the scene graph UI.

For now, state updates have to manually merge the data that needs to be updated.  But in a future change we may introduce updates based on the path of the data with a helper function that will automatically translate the update path to a newly updated state for a scene node.

NOTE:
Currently, nodes in the state manager and the scene manager are indexed by name. So if the name of a node is update in the scene graph UI, the dictionary of nodes will get out of sync and future updates for that node will break.  To fix that, we will need to allow to reindex the cache with the new name or use a immutable ID instead so that changing the name does not change how caches are indexed.