Changed the flow for calling functions to render a vertex buffer and shader program
Previously renderable classes would call `_bindShaderProgram` to set things up to render the vertex buffer with a shader program, which was a method on the base renderable class. And the renderable instance would call render function once the buffer and the shader program were bound.  Now there is a static function that takes a vertex buffer and a shader program, and it does the binding and calls the render function making all the steps involved for rendering something more clear.