Refactored resource-loader.js to support more file types
Currently the resource manager only supports loading obj files, so I made some choices based on that. But now Im planning on adding support for 3ds files so the time to refactor the resource manager to support loading other files has come.

The way Im approaching is that different file types will provice a loader to load files, and then use that data to update the node in the scene graph. This should allow for each file format to provide a loader, which will usually implement some worker and then the parser of the data specific for each file format.