Adding fbx binary file parser!
Context: currently scenic only support OBJ file formats, which do not support much besides geometry. But in looking to support animation, fbx is a good format. So the changes here are to add basic support for parsing fbx birnary files. Im starting with binary because that's the only format that blender can export, so it makes it easier to do work in blender and use directly in scenic without needing to convert from binary to text with a tool like "FBX Converter". Im starting with only supporting FBX version 7400, but in a follow up change I will add support for 7500.

There is a test for parsing a simple cube, which you can ru with `npx run test`.