Support for fbx connections and muliple models
The current support for fbx files is just for very simple rendering of geometry and there is no support for connections, which is how fbx defines relationships between items in the models in an fbx file. Currently, this adds support for Object to Object (OO) relationships.  For this to work, I create all the scene nodes for each type of object in the fbx and then just do a simple hierarchical mapping using the connections. The tedious is in creating the scene nodes.  I create a few artificial types of nodes to hold information, but the important ones are static mesh and group for hold geometry and model accordingly from an fbx file. Note that I have added support to read transform data so I have updated sphere.fbx to scale to 1 again because somehow blender will export models and apply a scale of 100.

Full list of connections:
OO: Object (source) to Object (destination).
OP: Object (source) to Property (destination).
PO: Property (source) to Object (destination).
PP: Property (source) to Property (destination).

You can see some details about connections in [1] under the Programmers guide.

1. https://download.autodesk.com/us/fbx/20112/fbx_sdk_help/index.html