Changed the state manager so that it can assign state to items added after the state manager is created.
Context: I am adding support for fbx file and to support models that have multiple parts I need to create a new hierarchy of scene nodes, most of which will need their state to be available in the state manager.  And that because these multi part fbx models are created from a single file reference in the scene config, but the scene config does not have their state so we need to create them at runtime.