Adding support to generate normal vectors for fbx files
fbx file support was only rendering simple geometry with no normal vector used for lighting. But in this diff Im adding the ability to build normal vectors. For this to work, I expand the indexed vertices to full triangles and use that to create all the normals for each polygon. This same logic is used for obj files so Im moving it to vector3.