Adding a few fbx models to show case fbx rendering
Now that we have support for more complex fbx files Im configuring the default scene to render a more complicated model. I am also reverted light spheres as obj files to keep that part simple for configuring the color of each sphere in the config file itself instead of the fbx file.