Adding fbx cube with normals and uv coordinates for better support of fbx files
Context: currently I am triangulating all the vertices from an fbx model and then building normals from those.  But ultimately we want to be able to load the pre-calculated normal vertices from the fbx geometry themselves. And we also want to add support for texturing. So in these changes Im just adding a new cube.fbx that has normal vertices and uv coordinates.