Correcting -0 test failure in matrix tests
A test is failing because -0 and 0 are techically different numbers, so strict equality fails.  However, in practice all math operations yield the same result. For example, multiplying any number times -0 will still yield -0. And all other operations behave the same.  So for now Im just tweaking the failing test and left a TODO to consider fixing mat4 to never generate -0.