Refactored ShaderProgram factory to better handle attributes and caching
The current implementation for the shader factory will cache the entire shader program which causes issues when the same needs to be used in a new program that isn't shared.  I have changed the caching logic to cache the source for the vertex and fragment shaders, and when creating a shader program I will compile and cache the compiled shader so that we can much more efficiently create new programs that essentially only need linking. Also cleaned up a few spots to make it more clear that we are not as eager to cache shader programs. The reason for this is that when rendering models that have multiple geometry groups, we want to use new programs that can be quickly created and reused for the same geometry (vertex buffer) so that once its bound to the buffer the program doesn't need to be bound again in every render call.

Another change is that previously we only support supported auto attribute binding for 3d and 4d vectors; there was no support for 2d vectors. We need support for 2d vectors for supporting UV coordinates for textures, which I am adding as part of supporting FBX files in follow up changes.

NOTE: a shader program is an instance of `ShaderProgram` and not a `WebGLProgram`.  A `ShaderProgram` is a wrapper for `WebGLProgram` tho.