Adding texture coordinates and refactored vertex buffer
VertexBuffer is the class that provides the function for rendering a buffer. And we have all sorts of other buffers that are wrappers to provide a bind function for WebGL, including VertexBufferData and VertexBufferIndexes. But they didn't provide clear semantics from the API they all provided because some of the buffers like VertexBufferIndexes and VertexBufferData both provided a render function as well as VertexBuffer.  The changes here are for removing the render method from VertexBufferIndexes and VertexBufferData and solely rely on VertexBuffer to provide the render function, which will choose what to do based on the buffers that are available.

Another tweak is to add TextureVertexBufferData to set us up for future work to support textures.