Adding getIndexed2DComponents for reading indexed data for 2d buffer like UV coordinates
So far everything we have been supporting is 3D data in 3D space. But as we are ramping for support for texturing, I am adding a new function getIndexed2DComponents to read UV indexed data to generate a new array with all the 2D information expanded out to make it more suitable for triangle rendering without the use of indexes.