Flipped texture rendering along the Y axis
Rendering textures that are loaded with new Image() are loaded up side down. So we flip that so that those images look correct.
As long as image buffer are stored in the same order as images loaded with new Image(), we can just keep this as default behavior. Otherwise, we are going to need to add a way to flip images more selectively on a per texture basis which can add lots of unwanted complexity because we will need to keep correctly push and pop that configuration from the state stack so that textures don't interfere with each other.

Helpful context: https://jameshfisher.com/2020/10/22/why-is-my-webgl-texture-upside-down/