Simplifying scene lighting in default scene
So far the scene has had several light sources with one that was animated rotating around the main object being rendered. This was primarily to show case frame animation and lightis.  But I am making that simpler by removing the animation and having a separate scene object group that so that interactions with the objects in the scene don't affect location/rotation of the default lights illuminating the entire system.  This will provide a more focused and simpler experience when rendering models, especially in the up coming FBX support with textures where the main thing is the fbx object and not any of the fancy lights in the scene.