Adding phong shader with texture support
In order to add support to textured objects, I need to update the phong shaders to support textures. These changes add a new vertex/fragment shader based on the phong-lighting shaders that add support for 6 textures that are all added together to produce the final texture value for a fragment. Yeah we need more fancy blending functions for textures, but this is will get us initial support and we will let us iterate on what we need.  Im also wiring up the new shader in the shader factory.