Refactored shader loader so that programs are created by the model loaders
Previously, the shader program factory would get a list of all the scene nodes and it would load the shaders dynamically and then configure each node.  This became a difficult workflow when looking to add more shaders and scene node types like what we are adding in FBX texturing support.  It also made the assumption that the shader was the same for a particular type of scene node (static-mesh was phong-lighting and light was flat-material), and that logic was hard coded in the shader factory causing lots of friction for supporting different file formats where static mesh nodes can have textures so phong-lighting is no longer the correct shader. Different file format also have the ability to specify their own shader, which is a limitation with the approach of having the shader factory figure out which shader is used.

In order to resolve all these issues, I am now preloading the known shaders from storage and caching them and letting the file format loaders pick and choose the correct shader!