Normal map support for FBX files
Phew! These changes add normal map support for bump lighting. The gist is that we use a normal map texture (texture with normal vectors) for pre-calculated normal vertors in each pixel in the texture. This gives us the ability to add tons of details to surfaces when calculating lighting.  This required a few things:

1. Build a TBN matrix. TBN stands for Tangent, Bitangent, and Normal vectors. This combination of vectors form a 3x3 matrix used for moving vectors between world space and tangent space.  This is important for bump lighting because normal map textures encode normal vectors in tangent space and light positions are in world space; the TBN matrix allows us to go between tangent space and world space.
2. Update shaders so that they correctly use the TBN matrix.  For this we have a couple of options.  1. Convert light positions to tangent space in the vertex shader so that the fragment shader has light positions in the correct space. TBN matrix multiplications are only per vertex. 2. Convert every normal vector to world space in the fragment shader. This option is much more expensive because TBN matrix multiplications are per pixel.  So I implemented the first option.  To implement the first option we simply multiply the light positions by the transpose of TBN to get the light positions in tangent space. We then use those light positions (which are in tangent space) in the fragment shader to calculate the dot product with the normal vectors from the normal map texture (normal vectors are already in tangent space).
3. Use the normal vector information in the normal map textures. This is the easiest part where we just read pixels from the normal map texture. One extra step in the process of using the texures is that textures data ranges from 0 to 1, but normal vectors range from -1 to 1. So we need to do a small extra step of converting from texture data to normal data with `normal * 2 - 1`.

Some of the tricky parts in implementing this were:
1. Correctly handle the fact that vertex vectors have three components and UV (texture) coordinates only have 2 components.  So the loop for iterating thru vertices to calculate TBN vectors need to correctly offset vertices and UV coordinates.  This actually lead to some nasty behavior where only a portion of the models were lit and other parts were not lit at all because some of the Tangets and Bitangents had `NaN`s in them. This was really hard to debug and ultimately more unit tests would have caught this.
2. Understand the implications of calculating TBN because there are a couple of approaches that you might find in your own research that are slgithly different, but yield the same results. The formula I used [2] gives us the Tanget and Bitangent by doing some linear transformations on the UV and vertices, which happens in javascript. Another approach here is to calculate the Tangent (in javascript) with a smaller version of the formula [2] and then do a cross product between the normal and the tangent vectors to get the bitangent in the vertex shader. You can use this latter approach if memory was a concern for storing yet another vertex buffer to hold that data.
3. Appreciating the fact that whichever approach you choose for calculating TBN that we understand that the Tangent, Bitangent, and Normal vectors are orthogonal.  This is important because to go between spaces (e.g. world and tangent space) you need the inverse matrix.  But orhtogonal matrices have a special property where the transpose will invert the matrix. This simplified vertex shaders quite a bit because I only have to call `transpose`, which as I just described will get you the matrix inverse which allows us to take the light position and move it to tangent space in the vertex shader.

Quick and over simplified context about lighting. Lighting (in a phong model) is mostly calculated with the dot product of the normal vector and light vector; there are other factors to take into account but this is what determines how a light source directly affects lighting on a surface. First things first. A normal vector is a vector that is perpendicular to a surface (in more general terms a plane), and a unit vector is a vector of length of 1.  The dotproduct of the normal vector and the light vector gives us information about the angle of incidence (the angle between the normal and the light vector) which determines how much a light source affects a surface. This is where things get confusing because we use the dot product between the normal vector and the light to determine information about how a light source affects a surface. If the dotproduct between a normal vector and a light vector is -1, it means the light is hitting the back of a surface so for all intents and purposes the surface is not lit. A 0 means the normal and light source are perpendicular (light and surface are parallel) so that part of the surface is not lit either. A 1 means the light source is parallel to the normal vector (perpendicular to the surface) so maximum lighting is applied. So really, any value between 0 and 1 means that a pixel is lit.

The dopdproduct range between -1 to 0 isn't particularly interesting in lighting because that's just means that a pixel isn't lit.  But it is important to ensure that the normals encoded in textures that go from 0 to 1 are converted to -1 to 1 so that normal vectors in textures that go from 0 to 0.5 can correctly reflect that a pixel is not lit. For example, a value of 0.2 from the texture directly used in lighting means a pixel is lit, but if 0.2 were converted to the range of -1 to 1, a 0.2 would then be -0.6, which means that the decoded normal vector value will yield no lighting.

There are lots of articles that explain this, but these ones were particularly helpful:
1. https://learnopengl.com/Advanced-Lighting/Normal-Mapping
2. https://www.gamasutra.com/view/feature/1515/messing_with_tangent_space.php?print=1.
3. https://math.hws.edu/graphicsbook/c3/s5.html.
4. https://math.hws.edu/graphicsbook/c7/s2.html.
5. https://www.sciencedirect.com/topics/computer-science/lighting-equation
6. https://en.wikipedia.org/wiki/Angle_of_incidence_(optics)
7. https://community.khronos.org/t/glsl-and-tbn-matrix-is-my-understanding-correct/61839