Refactored how scene nodes are created from fbx files
As Im looking to add support for animation, I need better support for `OP` connection types so that animation curve nodes can update the right nodes. In order to do that, I need access to the connection type as scene nodes are being created.  This change allowed me to have better handling for normal maps because I can now determine from the connection information between the texture and the material that a paticular texture is for normal mapping vs something else.  The traversal of fbx nodes is still breadth first to accurately determine how many textures a model has before we create the model node itself, which helps us determine what shader to use; if a model has no texture then we use a shader that does not use textures.