Revamped keyframe animation functionality
As Im working thru adding animation support for FBX files, I need to enhance the support for the keyframe animation utilities I already have.  So I have added three mathods for animation 2v, 3v, and scalars. I have also revamped how animation keyframe is calculate so that it is stateless and animation is always derrived at time progresses.  The current implementation does have the draw back that animation ought to start at 0 milliseconds for the animation to start from the beginning. This needs to be more flexible so that we can start an animation at any given time.