Refactored node update API
Context: Im currently working on adding animation support in the scene tree and the API for propagating rendering nodes does not provide all the information at time some nodes it.  For example, propagating animation matrix informaion which in this case needs to happen in the preRender cycle which does not provide animation timing.  So the changes here adds that whick also allowed us to move the projection calculations from the scene manager onto nodes that actually care about projection matrices.