Adding basic FBX animation support
These changes add basic support for keyframe animation for FBX files. This will load up all the animations that are defined in the `Connections` and will process the animation defined for each model. Right now I am just applying the first animation I find but later changes will add support for running a particular animation stack.

The way FBX animation works is that we have animation curve nodes which contains animation curves. These animation curves is where information about keyframes exist. And a model can have multiple of these animation curves configured in an animation stack [1]. An animation stack is one animation sequence or take [2], and these are the animation sequences available to be rendered.  So for now Im just doing the first animation associated with a model.

The only supported animation so far is for transformations; translation, rotation, and scaling. Animation sequences override whatever translation, rotation, and scaling information a model has. Meaning only transformation animation is used when rendering a model. The transforms are applied in the preRender traversal of the scene tree and a model basically will look for children nodes that are animation curve nodes calling `getValues` for each. The animation curves will in turn aggregate all the data from all its children animation curve and these values are what we use as the transform for the model. The animation curve children all use linear interpolation animation to generate smooth animations.

NOTE: for some reason I have to swap the Y/Z rotation data in animation. Im not certain, but this is likely an artifact of blender export.  I gotta look into this separately.  I have also left the code to build up the connections for all the animation stacks commented out for now, but I will enable once we are ready to wire animating stacks.  For now Im just picking whatever animation is associated with a model.

1. https://help.autodesk.com/view/FBX/2016/ENU/?guid=__files_GUID_9481A726_315C_4A58_A347_8AC95C2AF0F2_htm
2. https://help.autodesk.com/view/FBX/2016/ENU/?guid=__cpp_ref_class_fbx_take_info_html