Adding image loading cache for FBX files
FBX files can define image textures and so far they have been relying on the browser cache to ensure that we get cache images. However, the changes here are so taht we cache in JS the promise with the Image instance so that we dont have to even create the Image itself more than once.