Adding support for caching textures
Previously I had added support to cache images, which is helpful to make sure we are not loading the same image multiple times.  These changes now add support for caching texture node, which helps prevent loading the same image to the GPU more than once! This also improves load times quite a bit, especially for models that use the same texture in multiple models of the same fbx file.