Adding support for animation with time intervals
So far support for animation was based on the assumption that between each frame lasted exactly 1 second. But that was done for simplicity purposes to get smooth animation working. But in reality frames can last longer and sometimes even less than one second, so the changes here are to ensure that we can handle such cases. FBX files can often have such behavior where we might may one animation for translation will last 30 seconds and a rotation animation that also last 30 seconds but the first 15 seconds rotate and the last 15 it does not. The last 15 seconds of the rotation will besically be a 15 second long frame with no rotation.

FBX files have time interval encoded where `46186158000` is 1 second. So in the animation curve nodes I have to ensure that we do the proper scale translations between that and seconds that last 1000 milliseconds which is what requestAnimationFrame provides. Its a simple formula `46186158000*ms*0.001`; 0.001 is to convert milliseconds to seconds. So interpolation has to find the frame for a particular time based on the times frames that are provided, and we interpolate that keyframe values corresponding to that for the length of the frame.

Most things in animation remain the same and providing times for each frame is optional.