Refactor to use fbx connections for traversal and remove recursion during traversal
The changes are to clean up the mental model a lil bit between fnx node wrappers and connections by basing all the fbx node traversal on connections.  There is no more mixing in the traversal logic of fbxNodeWrapper as well as Connection.  Now everything is a connection.

The other change is that we are no longer using recursion to iterate over the fbx nodes; we just create a stack and keep iterating thru it.