Split up the scene rendering out into its own build
The editor app and the scene rendering and all the logic were all bundled together so every time that a change was made in the scene renderer all the JSX had to be rebuilt, which was slow.  Im splitting up the scene into its own bundle so that rebuild times are faster!