Added ability to use the currently select animation stack in FBX models
So far we have been just rendering the whatever animation data any animatable object had. But that doesn't work well when there are multiple animations for an FBX model since the animator will process everything related to a node. The changes here allow us to use the correct data for any animation stack.

Currently there is no UI for selecting an animation stack, so I have just exported the scene graph to the global object (window in the browser) so that we can manually configure per scene object the animation stack to use.  Future changes will allow selecting from the UI.