Renamed Model to Mesh in FBX scene-node
For a while now I have been calling meshes in fbx files `model` because when I first implemented fbx support I only supported meshes which are called `Model` in an FBX document. However, there are other types of `Model`. We have `Mesh` which contains geometry, textures, animation data and so on. We have `LimbNode` which is an animation bone. And we have a miscellaneous one `Null` that basically represents an armature in models with animation exported from blender.  For a lil while I have been using `Model`, `Armature`, and `Bone` but it is time to rename `Model` to `Mesh`.