Adding Animation scene node to encapsulate animation information
So far we ahve been adding animation node directly in the scene node for the fbx document which means that animation stacks existed along side the renderable node.  For better grouping I have added an `Animation` scene node which can be used to group animation stacks. This makes it more efficient to find animation information and will also make it easier to interact with animations as a whole when interfrating it into the editor tree because we wont have animation information scattered in a scene node's children.