Renamed byType to childrenByType in the base scene node
Each scene node has a collection of nodes which are also stored `byType`.  `byType` however isn't clear enough about what it really has in it, so I am renaming it to `childrenByType` to make sure it is clear that it is children nodes that are stored by type in that collection.