Adding unique name generation for Animation nodes and better animation selection handling
I ran into issues when a scene has multiple animations and we weren't passing into the animation stack selector the selection stored in the node state manager. So everytime a new animation was selected, the animation panel was not reselecting the current animation. The changes here ensure that we are correctly passing into the animation panel the selected stack from the node state.

This was further compounded by the fact that I had taken the shortcut to not have unique animation name, so Im also ensuring animation names are unique