Adding ability to enabled/disable normal smoothing in FBX files
By default I smooth out normal vectors when calculating the TBN matrix which is where we calculate normal vectors.  This is so that we can render cubes and other shapes that is not correct to smooth out normals because they are at 90 degree angle, and smoothing that will cause cubes and walls to not have edges.

TODO: we can be smarter and automatically figure out if shared vertices for normals pointing at a 90 degree angle. In which case we would disable smoothing.