Adding default texture in FBX meshes
If an FBX file does not have a texture configured, currently we will render the object completely black because there is no ambient color and no light reflection. And this is confusing because scenes can seem like nothing is rendering. So instead I am adding a default texture that allows us to at least render scene object so that you can see them.