Adding better handling of number inputs for XYZ coordinates
We used to not do much about ensuring proper format of numbers inserted in the Coordinates react component which would cause rendering errors when entirely emptying out the coordinate fields. The changes here ensure that we dont have leading 0 and or bad behavior when the coordinate fields are emptied out.

There is an interesting behavior that I will look closer at if it becomes too annoying.  But inserting an extra `-` in a field that already negative causes the input element to entirely clear out the input fields.