Moved setting transforms to scene nodes from the scene manager
Currently every node's world matrix gets initialized by the scene manager in the preRender phase if the node has the data for it. But there are node that will process these transforms themselves such as the case of animation. So we are essentially processing these transforms twice for every animated node. To avoid that I am moving the logic to set the transforms to the base scene node prerender which allows ingeriting nodes to override that behavior. The only thing to pay attention to is that inheriting for scene node and overriding preRender will require you to call `super.preRender`, which is will incurr a small mental load.